chatbot-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.28
langchain-groq
mcp>=1.1.0
pydantic>=2.9.0
python-dotenv>=1.0.1
redis>=5.0.0
upstash-redis>=0.15.0
slowapi>=0.1.9
sse-starlette>=2.1.0

